Add TeamBuilder.Models
