+++ TeamBuilder.API/DataLoader/TeamByIdDataLoader.cs
@@ -1,11 +1,12 @@
+using GreenDonut;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore;
 using TeamBuilder.API.Data;
-using GreenDonut;
+using TeamBuilder.Models;
 
 namespace TeamBuilder.API.DataLoader
 {
