+++ TeamBuilder.API/TeamMembers/TeamMemberMutations.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using TeamBuilder.API.Data;
 using TeamBuilder.API.Services;
+using TeamBuilder.Models;
 
 namespace TeamBuilder.API.TeamMembers
 {
