+++ TeamBuilder.API/TeamMembers/TeamMemberPayloadBase.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using TeamBuilder.API.Common;
-using TeamBuilder.API.Data;
+using TeamBuilder.Models;
 
 namespace TeamBuilder.API.TeamMembers
 {
