+++ TeamBuilder.API/Teams/TeamMutations.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using TeamBuilder.API.Common;
 using TeamBuilder.API.Data;
 using TeamBuilder.API.Services;
+using TeamBuilder.Models;
 
 namespace TeamBuilder.API.Teams
 {
