+++ TeamBuilder.API/Types/TeamMemberType.cs
@@ -2,8 +2,8 @@
 using HotChocolate.Types;
 using System.Threading;
 using System.Threading.Tasks;
-using TeamBuilder.API.Data;
 using TeamBuilder.API.DataLoader;
+using TeamBuilder.Models;
 
 namespace TeamBuilder.API.Types
 {
