+++ TeamBuilder.GitHub/TeamBuilder.GitHub.csproj
@@ -8,15 +8,16 @@
   </PropertyGroup>
 
   <ItemGroup>
-  </ItemGroup>
-
-  <ItemGroup>
 	  <PackageReference Include="Azure.Identity" Version="1.5.0" />
 	  <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.3.0" />
 	  <PackageReference Include="GitHubJwt" Version="0.0.5" />
 	  <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.3" />
 	  <PackageReference Include="Octokit" Version="0.50.0" />
 	  <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\TeamBuilder.Models\TeamBuilder.Models.csproj" />
   </ItemGroup>
 
 </Project>
