+++ TeamBuilder.Models/ChallengeArea.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 
-namespace TeamBuilder.API.Data
+namespace TeamBuilder.Models
 {
     public class ChallengeArea
     {
