+++ TeamBuilder.Models/Team.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace TeamBuilder.API.Data
+namespace TeamBuilder.Models
 {
     public class Team
     {
