+++ TeamBuilder.Models/TeamBuilder.Models.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<TargetFramework>net5.0</TargetFramework>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Microsoft.Graph" Version="4.13.0" />
+	</ItemGroup>
+
+</Project>
