+++ TeamBuilder.Serverless/Functions/AddGitHubRepository.cs
@@ -5,14 +5,15 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.EventGrid;
 using Microsoft.Extensions.Logging;
 
-namespace TeamBuilder.Serverless
+namespace TeamBuilder.Serverless.Functions;
+
+public static class AddGitHubRepository
 {
-    public static class AddGitHubRepository
+    [FunctionName("AddGitHubRepository")]
+    public static void Run([EventGridTrigger]EventGridEvent eventGridEvent, ILogger log)
     {
-        [FunctionName("AddGitHubRepository")]
-        public static void Run([EventGridTrigger]EventGridEvent eventGridEvent, ILogger log)
-        {
-            log.LogInformation(eventGridEvent.Data.ToString());
-        }
+        log.LogInformation(eventGridEvent.Data.ToString());
+
+        //eventGridEvent.Data as Team
     }
 }
